chore(deps): neoforge-1.21.5
- Updated mod version from 1.21.4-1.12.0-beta.1 to 1.21.5-1.12.0-beta.1
- Updated Minecraft version from 1.21.4 to 1.21.5
- Updated NeoForge version from 21.4.155 to 21.5.95
- Adjusted version ranges in mod metadata files
- Fixed fall distance casting in TerrestrialStepSoundGenerator
- Removed unused ArmorItem import
- Simplified getUniqueFileName method by removing extension parameter

+++ src/main/java/eu/ha3/presencefootsteps/sound/generator/TerrestrialStepSoundGenerator.java
@@ -26,7 +26,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ArmorItem;
+
 
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
@@ -106,7 +106,7 @@ class TerrestrialStepSoundGenerator implements StepSoundGenerator {
         simulateAirborne();
         simulateBrushes();
         simulateStationary();
-        lastFallDistance = entity.fallDistance;
+        lastFallDistance = (float) entity.fallDistance;
     }
 
     protected void simulateStationary() {
